Fixes[MemberDetail] : Refactor CSS files in src/screens/MemberDetail and updated the import file (#3373)
* Refactor css

* Update customTableCell.tsx

* Update customTableCell.tsx
